Fixed some warning indicated on Solaris: cast return value of getuid() for printf in xmain.c; fixed ambiguous 'else' in unithand.c From a patch submitted by Fabrice Noilhan <Fabrice.Noilhan@ens.fr>.
[[originally from CVS; cvs2svn created svn r1161]]

+++ client/xmain.c
@@ -318,7 +318,7 @@ void x_main(int argc, char *argv[])
     else {
       freelog(LOG_NORMAL, "Your getpwuid call failed.  Please report this.");
       strcpy(name, "operator 00000");
-      sprintf(name+9, "%05i",getuid());
+      sprintf(name+9, "%05i", (int)getuid());
     }
   }
